Add more NAT-T support, don't send id to kernel...
to work around a bug in how the src and dst id
is determined. When we send the kernel the id, pfkey
reports that it receive an invalid value. Now, the
kernel will install security associations and policies
but still traffic does not flow between the peers.

Always use ipsec level require